Added entrypoint.rs
